Fix scrolling jumps
libinput currently scales wheel events by DEFAULT_AXIS_STEP_DISTANCE,
which is currently 10.
We may be able to find the value automatically, but it would require
some dark codings that may not be robust.
For the time being, just duplicate the define in libinput.c and remember
to update it if it is changed in the future.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>