Expose properties for the various configuration settings.
The property support isn't ideal, the server relies on the driver to check
first if a setting is possible, applying the setting itself cannot fail.
Thus we can't just check the libinput config status code, which matters mostly
in the sendevents property where we simply check that only a single bit is set
and that the supported modes are equal or more to the requested mode.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>